pyproject.toml
[tool.poetry]
name = "machine-tools"
version = "0.2.34"
description = "Пакет для работы с базой данных станков"
authors = ["Andrey Nikolaevich Korenyuk <korenyuk.a.n@mail.ru>"]
readme = "README.md"
packages = [{include = "machine_tools"}]

[tool.poetry.dependencies]
python = "^3.9"
dependency-injector = "^4.41.0"
pydantic = "^2.11.3"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
psycopg2-binary = "^2.9.10"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
mypy = "^1.5.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
machine_tools = "machine_tools.cli:main"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
skip-string-normalization = true
extend-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=machine_tools --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true